Put version-related methods in version utils class
